menambahkan crud pembayaran dan dashboard keuangan

+++ application/controllers/Admin.php
@@ -7,9 +7,9 @@ class admin extends CI_Controller
     function __construct()
     {
         parent::  __construct();
-  $this->load->model('m_model');
-  $this->load->helper('my_helper');;
-  $this->load->library('upload');
+     $this->load->model('m_model');
+     $this->load->helper('my_helper');
+     $this->load->library('upload');
         if ($this->session->userdata('logged_in')!=true) {
             redirect(base_url().'auth');
         }  
@@ -32,20 +32,19 @@ class admin extends CI_Controller
         $config['allowed_types'] = 'jpg|png|jpeg';
         $config['max_size'] = '30000';
         $config['file_name'] = $kode;
-        $this->upload->initialize($config);
-        if (!$this->upload->do_upload($value))
-        {
-            return array( false, '' );
-        }
-        else
-        {
-            $fn = $this->upload->$data();
+        
+        $this->load->library('upload', $config); // Load library 'upload' with config
+        
+        if (!$this->upload->do_upload($value)) {
+            return array(false, '');
+        } else {
+            $fn = $this->upload->data();
             $nama = $fn['file_name'];
-            return array( true, $nama );
-           
-
+            return array(true, $nama);
         }
     }
+
+    
 // DATASISWA
     public function datasiswa()
     {
@@ -59,53 +58,85 @@ class admin extends CI_Controller
     }
    
     public function aksi_tambahsiswa()
-    {
-        $foto=$this->upload_img('foto');
-       if ($foto[0] == false) {
-        $data=[
-            'foto' => 'User.png',
-            'nama_siswa' => $this->input->post('nama'),
-            'nisn' => $this->input->post('nisn'),
-            'gender' => $this->input->post('gender'),
-            'id_kelas' => $this->input->post('id_kelas'),
-        ];
-        $this->m_model->tambah_data('siswa', $data);
-        redirect(base_url('admin/datasiswa'));
-       } else {
-        $data=[
-            'foto' => $foto[1],
-            'nama_siswa' => $this->input->post('nama'),
-            'nisn' => $this->input->post('nisn'),
-            'gender' => $this->input->post('gender'),
-            'id_kelas' => $this->input->post('id_kelas'),
-        ];
-        $this->m_model->tambah_data('siswa', $data);
-        redirect(base_url('admin/datasiswa'));
-       }
-       
-    }
-    public function aksi_ubah_siswa()
-    {
-        $data = array (
-            'nama_siswa'=> $this->input->post('nama'),
-            'nisn'=> $this->input->post('nisn'),
-            'gender'=> $this->input->post('gender'),
-            'id_kelas'=> $this->input->post('id_kelas'),
-        );
-        $eksekusi=$this->m_model->ubah_data
-        ('siswa',$data,array('id_siswa'=>$this->input->post('id_siswa')));
-        if($eksekusi)
-        {
-            $this->session->set_flashdata('sukses','berhasil');
-            redirect(base_url('admin/datasiswa'));
-        }
-        else
-        {
-            $this->session->set_flashdata('error','gagal...');
-            redirect(base_url('admin/datasiswa/ubah_siswa/'.$this->input->post('id_siswa')));
+  {
+    $file_name = $_FILES['foto']['name'];
+    $file_temp = $_FILES['foto']['tmp_name'];
+    $kode = round(microtime(true) * 1000);
+    $file_name = $kode . '_' . $file_name;
+    $upload_path = './images/siswa/' . $file_name;
 
-        }
+    if (move_uploaded_file($file_temp, $upload_path)) {
+      $data = [
+        'foto' => $file_name,
+        'nama_siswa' => $this->input->post('nama'),
+        'nisn' => $this->input->post('nisn'),
+        'gender' => $this->input->post('gender'),
+        'id_kelas' => $this->input->post('id_kelas'),
+      ];
+      $this->m_model->tambah_data('siswa', $data);
+      redirect(base_url('admin/datasiswa'));
+    } else {
+      $data = [
+        'foto' => 'User.png',
+        'nama_siswa' => $this->input->post('nama'),
+        'nisn' => $this->input->post('nisn'),
+        'gender' => $this->input->post('gender'),
+        'id_kelas' => $this->input->post('id_kelas'),
+      ];
+      $this->m_model->tambah_data('siswa', $data);
+      redirect(base_url('admin/datasiswa'));
     }
+  }
+
+  public function aksi_ubah_siswa()
+	{
+		$foto = $_FILES['foto']['name'];
+		$foto_temp = $_FILES['foto']['tmp_name'];
+
+		// Jika ada foto yang diunggah
+		if ($foto) {
+			$kode = round(microtime(true) * 1000);
+			$file_name = $kode . '_' . $foto;
+			$upload_path = './images/siswa/' . $file_name;
+
+			if (move_uploaded_file($foto_temp, $upload_path)) {
+				// Hapus foto lama jika ada
+				$old_file = $this->m_model->get_siswa_foto_by_id($this->input->post('id_siswa'));
+				if ($old_file && file_exists('./images/siswa/' . $old_file)) {
+					unlink('./images/siswa/' . $old_file);
+				}
+
+				$data = [
+					'foto' => $file_name,
+					'nama_siswa' => $this->input->post('nama'),
+					'nisn' => $this->input->post('nisn'),
+					'gender' => $this->input->post('gender'),
+					'id_kelas' => $this->input->post('id_kelas'),
+				];
+			} else {
+				// Gagal mengunggah foto baru
+				redirect(base_url('admin/ubah_siswa/' . $this->input->post('id_siswa')));
+			}
+		} else {
+			// Jika tidak ada foto yang diunggah
+			$data = [
+				'nama_siswa' => $this->input->post('nama'),
+				'nisn' => $this->input->post('nisn'),
+				'gender' => $this->input->post('gender'),
+				'id_kelas' => $this->input->post('id_kelas'),
+			];
+		}
+
+		// Eksekusi dengan model ubah_data
+		$eksekusi = $this->m_model->ubah_data('siswa', $data, array('id_siswa' => $this->input->post('id_siswa')));
+
+		if ($eksekusi) {
+			redirect(base_url('admin/datasiswa'));
+		} else {
+			redirect(base_url('admin/ubah_siswa/' . $this->input->post('id_siswa')));
+		}
+	}
+
     public function ubah_siswa($id) 
     {
         $data['siswa']=$this->m_model->get_by_id('siswa', 'id_siswa', $id)->result();
@@ -114,11 +145,42 @@ class admin extends CI_Controller
     }
     
 
+    // public function hapus_siswa($id)
+    // {
+    //    $this->m_model->delete('siswa', 'id_siswa', $id);
+    //     redirect(base_url('admin/datasiswa'));
+    // }
+
     public function hapus_siswa($id)
     {
-       $this->m_model->delete('siswa', 'id_siswa', $id);
-        redirect(base_url('admin/datasiswa'));
+       $siswa=$this->m_model->get_by_id('siswa', 'id_siswa', $id)->row();
+       if ($siswa) {
+           if ($siswa->foto !== 'User.png') {
+               $file_path = '.images/siswa/' . $siswa->foto;
+
+               if (file_exists($file_path)) {
+                   if (unlink($file_path)) {
+                       // hapus file berhasil menggunakan model delete
+                       $this->m_model->delete('siswa', 'id_siswa', $id);
+                       redirect(base_url('admin/datasiswa'));
+                    } else {
+                         // gagal menghapus file
+                        echo "gagal menghapus file.";
+                    }
+                } else {
+                     // file tidak ditemukan
+                     echo "file tidak ditemukan.";
+                }  
+            } else {
+                 // jangan hapus file 'User.png'
+                 $this->m_model->delete('siswa', 'id_siswa', $id);
+                redirect(base_url('admin/datasiswa'));
+            }
+    } else {
+        // siswa tidak ditemukan
+        echo "siswa tidak ditemukan.";
     }
+}
 
 // AKUN
     public function akun()
